Train backward changed: swap things *in memory*.
The dataset OpenSubData has duplicate t_given_s and s_given_t.
They are supposed to be:

t_s: 1 2 3|4 5 6
s_t: 4 5 6|1 2 3

but, actually:
t_s: 1 2 3|4 5 6
s_t: 1 2 3|4 5 6

We will take more cpu time to correct this.
And two files just differ in their order can bring a lot of trouble:
1. waste space.
2. cause inconsistency in train_test split. Since we random shuffle things before a split.
the pairs after the split may not even contain the same examples!
3. If you use the wrong files, you train the wrong models.

We come up with a nice fix: tell the Dataset to swap things itself on the fly when loaded.
This fixes all troubles above.
It isn't that nice?